+++ src/redux/initialStore.js
@@ -1,5 +1,8 @@
-import {createStore} from "redux";
-import {messageReducer} from "./reducers/message";
+import {combineReducers, createStore} from "redux";
+import {chatReducer} from "./reducers/chats";
+import {messageReducer} from "./reducers/messages";
 
- export const InitialStore = createStore(messageReducer);
+const reducer = combineReducers(chatReducer, messageReducer);
+
+ export const InitialStore = createStore(reducer);
